Test Automation for LearnAbout. Updating to be in synch with other test automation changes made.
