+++ test/src/org/labkey/test/pages/cds/XAxisVariableSelector.java
@@ -68,7 +68,7 @@ public class XAxisVariableSelector extends DataspaceVariableSelector
         if (_test.isElementPresent(Locator.tagWithClass("button", "yaxisbtn").notHidden()))
         {
             _test.sleep(1500);
-            _test._ext4Helper.waitForMaskToDisappear(60000);
+            _test._ext4Helper.waitForMaskToDisappear(120000); // Wait 2 mins. The test have much lower performance on TC. Until we have a real performance test (consistent environment etc...) I would rather not fail function test for it.
             // There is a bug where the mouse can end up over a time axis data point which will generate a hopscotch bubble.
             // However that is not the bubble indicating median values. So moving mouse out of the way.
             _test.mouseOver(Locator.xpath("//img[contains(@src, 'logo.png')]"));
