+++ test/src/org/labkey/test/pages/cds/YAxisVariableSelector.java
@@ -72,7 +72,7 @@ public class YAxisVariableSelector extends DataspaceVariableSelector
         _test.click(CDSHelper.Locators.cdsButtonLocator("Set y-axis"));
         // This code should wait for the plot to draw and if there is a 'Median values' pop-up wait for it to go away as well.
         _test.sleep(1500);
-        _test._ext4Helper.waitForMaskToDisappear(60000);
+        _test._ext4Helper.waitForMaskToDisappear(120000);  // Wait 2 mins. The test have much lower performance on TC. Until we have a real performance test (consistent environment etc...) I would rather not fail function test for it.
         // There is a bug where the mouse can end up over a time axis data point which will generate a hopscotch bubble.
         // However that is not the bubble indicating median values. So moving mouse out of the way.
         _test.mouseOver(Locator.xpath("//img[contains(@src, 'logo.png')]"));
